Logging clean-up

+++ src/main/java/dk/purplegreen/musiclibraryboot/service/MusicLibraryService.java
@@ -35,17 +35,16 @@ public class MusicLibraryService {
 
 	@Transactional(readOnly = true)
 	public List<Artist> getArtists() {
-		if (log.isDebugEnabled()) {
-			log.debug("getArtists() called");
-		}
+
+		log.debug("getArtists() called");
+
 		return artistRepository.findAllByOrderByNameAsc();
 	}
 
 	@Transactional(readOnly = true)
 	public Artist getArtist(Integer id) throws MusicLibraryServiceException {
-		if (log.isDebugEnabled()) {
-			log.debug("getArtist() called with id: " + id);
-		}
+
+		log.debug("getArtist() called with id: {}", id);
 
 		Artist result = artistRepository.findOne(id);
 		if (result == null)
@@ -56,12 +55,10 @@ public class MusicLibraryService {
 	@Transactional
 	public Artist createArtist(Artist artist) throws MusicLibraryServiceException {
 
+		log.debug("createArtist() called with artist: {}", artist);
+
 		if (artist == null) {
 			throw new InvalidArtistException("Artist cannot be null");
-		}
-
-		if (log.isDebugEnabled()) {
-			log.debug("createArtist called with artist: " + artist.getName());
 		}
 
 		artist.setId(null);
@@ -72,12 +69,10 @@ public class MusicLibraryService {
 	@Transactional
 	public Artist updateArtist(Artist artist) throws MusicLibraryServiceException {
 
+		log.debug("updateArtist() called with artist: {}", artist);
+
 		if (artist == null) {
 			throw new InvalidArtistException("Artist cannot be null");
-		}
-
-		if (log.isDebugEnabled()) {
-			log.debug("updateArtist called with artist: " + artist.getId() + "-" + artist.getName());
 		}
 
 		getArtist(artist.getId());
@@ -88,35 +83,35 @@ public class MusicLibraryService {
 	@Transactional
 	public void deleteArtist(Artist artist) throws MusicLibraryServiceException {
 
+		log.debug("deleteArtist() called with artist: {}", artist);
+
+		if (artist == null) {
+			throw new InvalidArtistException("Artist cannot be null");
+		}
+
 		artist = getArtist(artist.getId());
 
 		if (albumRepository.countByArtist(artist) > 0) {
 			throw new InvalidArtistException("Artist has albums and cannot be deleted");
 		}
 
-		if (log.isDebugEnabled()) {
-			log.debug("deleteArtist() called with id: " + artist.getId());
-		}
 		artistRepository.delete(artist);
 	}
 
 	@Transactional(readOnly = true)
 	public List<Album> getAlbums(Artist artist) throws MusicLibraryServiceException {
 
+		log.debug("getAlbums() called for artist: {}", artist);
+
 		artist = getArtist(artist.getId());
 
-		if (log.isDebugEnabled()) {
-			log.debug("getAlbums() called for artist: " + artist.getName());
-		}
 		return albumRepository.findByArtist(artist);
 	}
 
 	@Transactional(readOnly = true)
 	public List<Album> findAlbums(String artist, String title, Integer year) {
 
-		if (log.isDebugEnabled()) {
-			log.debug("findAlbums called: artist=" + artist + ", title=" + title + ", year=" + year);
-		}
+		log.debug("findAlbums() called: artist={}, title={}, year={}", artist, title, year);
 
 		List<Album> result;
 
@@ -132,9 +127,7 @@ public class MusicLibraryService {
 	@Transactional(readOnly = true)
 	public Album getAlbum(Integer id) throws MusicLibraryServiceException {
 
-		if (log.isDebugEnabled()) {
-			log.debug("getAlbum() called with id: " + id);
-		}
+		log.debug("getAlbum() called with id: {}", id);
 
 		Album result = albumRepository.findOne(id);
 		if (result == null)
@@ -144,9 +137,8 @@ public class MusicLibraryService {
 
 	@Transactional(readOnly = true)
 	public List<Album> getAlbums() {
-		if (log.isDebugEnabled()) {
-			log.debug("getAlbums() called");
-		}
+
+		log.debug("getAlbums() called");
 
 		return albumRepository.findAllByOrderByTitleAsc();
 	}
@@ -154,12 +146,10 @@ public class MusicLibraryService {
 	@Transactional
 	public Album createAlbum(Album album) throws MusicLibraryServiceException {
 
+		log.debug("createAlbum() called with album: {}", album);
+
 		if (album == null) {
 			throw new InvalidAlbumException("Album cannot be null");
-		}
-
-		if (log.isDebugEnabled()) {
-			log.debug("createAlbum called with artist: " + album.getTitle());
 		}
 
 		album.setId(null);
@@ -171,12 +161,10 @@ public class MusicLibraryService {
 	@Transactional
 	public Album updateAlbum(Album album) throws MusicLibraryServiceException {
 
+		log.debug("updateAlbum() called with album: {}", album);
+
 		if (album == null) {
 			throw new InvalidAlbumException("Album cannot be null");
-		}
-
-		if (log.isDebugEnabled()) {
-			log.debug("updateAlbum called with album: " + album.getId() + "-" + album.getTitle());
 		}
 
 		getAlbum(album.getId());
@@ -187,13 +175,13 @@ public class MusicLibraryService {
 
 	@Transactional
 	public void deleteAlbum(Album album) throws MusicLibraryServiceException {
-		if (log.isDebugEnabled()) {
-			log.debug("deleteAlbum() called with id: " + album.getId());
-		}
+
+		log.debug("deleteAlbum() called with album: {}", album);
+
 		albumRepository.delete(getAlbum(album.getId()));
 	}
 
 	private void attachSongs(Album album) {
-		album.getSongs().stream().forEach(song -> song.setAlbum(album));			
+		album.getSongs().stream().forEach(song -> song.setAlbum(album));
 	}
 }
